StatTool2: int->size_t return type of perfmeta_get_values_memsize

+++ src/base/perf_metadata.h
@@ -599,7 +599,7 @@ void perfmeta_custom_dump_stats_in_table_form (const UINT64 ** stats, int no_of_
 void perfmeta_copy_stats (UINT64 * dst, UINT64 * src);
 void perfmeta_get_stat_index_and_dimension (const char *stat_name, const char *dimension_name, int *stat_index,
 					    int *fixed_dimension);
-extern int perfmeta_get_values_memsize (void);
+extern size_t perfmeta_get_values_memsize (void);
 extern int perfmeta_get_values_count (void);
 
 #endif /*_PERF_METADATA_H_*/
